refactor: 회원 관리 코드 분리
